+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/states/connecting_state.rs
@@ -97,7 +97,7 @@ impl ConnectingState {
             let mut bridge_endpoints = Vec::new();
             if shared_state.tunnel_settings.bridges_enabled()
                 && let Some(gateways) = &selected_gateways
-                && let Some(params) = &gateways.entry.bridge_params
+                && let Some(params) = &gateways.entry_gateway().bridge_params
             {
                 bridge_endpoints = params.get_addrs();
             }
@@ -109,7 +109,7 @@ impl ConnectingState {
                 bridge_endpoints,
                 ws_entry_endpoints: selected_gateways
                     .as_ref()
-                    .map(|v| v.entry.endpoints())
+                    .map(|v| v.entry_gateway().endpoints())
                     .unwrap_or_default(),
                 api_endpoints: Vec::new(),
                 dns_servers: shared_state.tunnel_settings.default_dns_ips(),
@@ -144,8 +144,8 @@ impl ConnectingState {
             selected_gateways
                 .as_ref()
                 .map(|gateways| EstablishConnectionData {
-                    entry_gateway: GatewayId::from(*gateways.entry.clone()),
-                    exit_gateway: GatewayId::from(*gateways.exit.clone()),
+                    entry_gateway: GatewayId::from(gateways.entry_gateway().clone()),
+                    exit_gateway: GatewayId::from(gateways.exit_gateway().clone()),
                     tunnel: None,
                 });
 
@@ -331,6 +331,7 @@ impl ConnectingState {
             shared_state.gateway_cache_handle.clone(),
             shared_state.topology_provider.clone(),
             tunnel_monitor_event_sender,
+            shared_state.wg_keys_db.clone(),
             #[cfg(not(any(target_os = "android", target_os = "ios")))]
             shared_state.route_handler.clone(),
             #[cfg(any(target_os = "ios", target_os = "android"))]
@@ -367,12 +368,12 @@ impl ConnectingState {
         #[cfg(not(any(target_os = "android", target_os = "ios")))]
         let set_policy_result = {
             if _shared_state.tunnel_settings.bridges_enabled()
-                && let Some(params) = &gateways.entry.bridge_params
+                && let Some(params) = &gateways.entry_gateway().bridge_params
             {
                 self.firewall_policy_params.bridge_endpoints = params.get_addrs()
             }
 
-            self.firewall_policy_params.ws_entry_endpoints = gateways.entry.endpoints();
+            self.firewall_policy_params.ws_entry_endpoints = gateways.entry_gateway().endpoints();
             Self::set_firewall_policy(_shared_state, &self.firewall_policy_params)
         };
         self.selected_gateways = Some(*gateways);