+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel/mod.rs
@@ -10,6 +10,7 @@ pub mod wireguard;
 
 pub use gateway_selector::SelectedGateways;
 use nym_gateway_directory::GatewayCacheHandle;
+use nym_vpn_store::keys::wireguard::WireguardKeysDb;
 use tokio_util::sync::CancellationToken;
 
 #[cfg(windows)]
@@ -21,10 +22,11 @@ pub use tombstone::Tombstone;
 pub async fn select_gateways(
     gateway_cache_handle: GatewayCacheHandle,
     tunnel_settings: &TunnelSettings,
+    wg_keys_db: WireguardKeysDb,
     cancel_token: CancellationToken,
 ) -> Result<SelectedGateways> {
     let select_gateways_fut =
-        gateway_selector::select_gateways(gateway_cache_handle, tunnel_settings);
+        gateway_selector::select_gateways(gateway_cache_handle, tunnel_settings, wg_keys_db);
 
     cancel_token
         .run_until_cancelled(select_gateways_fut)