+++ nym-vpn-core/crates/nym-vpn-store/src/keys/wireguard/persistence/ephemeral.rs
@@ -1,16 +1,16 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use std::collections::HashMap;
+use std::{collections::HashMap, sync::Arc};
 
 use tokio::sync::Mutex;
 
 use crate::keys::wireguard::{WireguardKeyStore, WireguardKeys, persistence::random_keys};
 
 #[allow(dead_code)]
-#[derive(Default)]
+#[derive(Clone, Default)]
 pub struct InMemEphemeralKeys {
-    keys: Mutex<HashMap<String, WireguardKeys>>,
+    keys: Arc<Mutex<HashMap<String, WireguardKeys>>>,
 }
 
 #[derive(Debug, thiserror::Error)]